requirements.txt
aiogram>=3.0,<4.0
aiohttp>=3.8,<4.0
beautifulsoup4>=4.12
python-dotenv>=1.0
